add duty-cycle record-64 bits.

+++ sys/include/net/gnrc/gomach/types.h
@@ -289,10 +289,10 @@ typedef struct gomach {
 
 #if (GNRC_GOMACH_ENABLE_DUTYCYLE_RECORD == 1)
     /* Parameters for recording duty-cycle */
-    uint32_t last_radio_on_time_ticks;                          /**< The last time in ticks when radio is on */
-    uint32_t radio_off_time_ticks;                              /**< The time in ticks when radio is off */
-    uint32_t system_start_time_ticks;                           /**< The time in ticks when chip is started */
-    uint32_t awake_duration_sum_ticks;                          /**< The sum of time in ticks when radio is on */
+    uint64_t last_radio_on_time_ticks;                          /**< The last time in ticks when radio is on */
+    uint64_t radio_off_time_ticks;                              /**< The time in ticks when radio is off */
+    uint64_t system_start_time_ticks;                           /**< The time in ticks when chip is started */
+    uint64_t awake_duration_sum_ticks;                          /**< The sum of time in ticks when radio is on */
 #endif
 } gnrc_gomach_t;
